Utils.Drasil

Documentation

enumBullet :: Reference -> [Sentence] -> LabelledContent

enumBullet applies Enumeration, Bullet and Flat to a list

enumBulletU :: [Sentence] -> Contents

same as enumBullet but unlabelled

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

enumSimple enumerates a list and applies Simple and Enumeration to it: s - start index for the enumeration, t - title of the list, l - list to be enumerated,

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

same as enumSimple but unlabelled

eqUnR :: Expr -> Reference -> LabelledContent

Constructs LabelledContent from an expression and a reference

eqUnR' :: Expr -> Contents

Same as eqUnR except content is unlabelled

mkEnumSimpleD :: (Referable c, HasShortName c, Definition c) => [c] -> [Contents]

mkEnumSimpleD is a convenience function for transforming types which are instances of the constraints Referable, HasShortName, and Definition, into Simple-type Enumerations.

blank :: Doc

indent :: Doc -> Doc

indentList :: [Doc] -> Doc

capitalize :: String -> String

String capitalization

stringList :: [String] -> String

Comma separated list with "and" before final item

data EnumType

Constructors

Numb

Upper

Lower

data WrapType

Constructors

Parens

Period

data SepType

Constructors

Comma

SemiCol

data FoldType

Constructors

List

Options

foldConstraints :: Quantity c => c -> [Constraint] -> Sentence

Helper for formatting constraints

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements with "enumerators" in "wrappers" using foldlList

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements separated by a "separator", ending with "and" or "or"

foldlSP :: [Sentence] -> Contents

Fold sentences then turns into content using foldlSent

foldlSP_ :: [Sentence] -> Contents

Same as foldlSP but uses foldlSent_

foldlSPCol :: [Sentence] -> Contents

Same as foldlSP but uses foldlSentCol

foldlSent :: [Sentence] -> Sentence

Partial function application of foldle for sentences specifically. folds with spaces and adds "." at the end.

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon

foldlsC :: [Sentence] -> Sentence

Folds a list of elements separated by commas, including the last element

foldNums :: String -> [Int] -> Sentence

Parses a list of integers into a nice sentence (ie. S "1, 4-7, and 13")

numList :: String -> [Int] -> [String]

Parses a list of integers into a list of strings (ie. ["1", "4-7", "13"])

addPercent :: Show a => a -> Sentence

Converts input to a Sentence and appends %

bulletFlat :: [Sentence] -> ListType

Applies Bullet and Flat to a list.

bulletNested :: [Sentence] -> [ListType] -> ListType

Applies Bullets and headers to a Nested ListType: t - Headers of the Nested lists, l - Lists of ListType

checkValidStr :: String -> String -> Either String String

Uses an Either type to check if a string is valid - Left with error message if invalid Char in String, else Right with String

chgsStart :: (HasShortName x, Referable x) => x -> Sentence -> Sentence

Output is of the form "reference - sentence"

definedIn :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence

Takes a HasSymbol that is also Referable and outputs as a Sentence "symbol is defined in reference"

definedIn' :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence -> Sentence

Same as definedIn, but allows for addition information to be appended

definedIn'' :: (Referable r, HasShortName r) => r -> Sentence

Takes a Referable and outputs as a Sentence "defined in reference" (no HasSymbol)

displayStrConstrntsAsSet :: Quantity a => a -> [String] -> Sentence

Produces a sentence that displays the constraints in a {}.

displayDblConstrntsAsSet :: Quantity a => a -> [Double] -> Sentence

Produces a sentence that displays the constraints in a {}.

eqN :: Int -> Sentence

Prepends the word Equation to an Int

eqnWSource :: (Referable r, HasShortName r) => Expr -> r -> Sentence

Takes an expression and a referable and outputs as a Sentence "expression (source)"

fromReplace :: (Referable r, HasShortName r) => r -> UnitalChunk -> Sentence

Takes a referable source and a UnitalChunk and outputs as a Sentence "From source we can replace symbol"

fromSource :: (Referable r, HasShortName r) => r -> Sentence

Wraps "from reference" in parentheses

fromSources :: (Referable r, HasShortName r) => [r] -> Sentence

Similar to fromSource but takes a list of references

fmtU :: MayHaveUnit a => Sentence -> a -> Sentence

Takes an amount and adds a unit to it: n - sentenc representing an amount, u - unit we want to attach to amount

follows :: (Referable r, HasShortName r) => Sentence -> r -> Sentence

Appends "following reference" to the end of a Sentence

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in..."

itemRefToSent :: String -> Sentence -> Sentence

Makes Sentences from an item and its reference: a - String title of reference, b - Sentence containing the full reference

makeListRef :: [a] -> Section -> [Sentence]

Takes a list and a reference, generates references to match the length of the list: l - list whos length is to be matched

makeTMatrix :: Eq a => [Sentence] -> [[a]] -> [a] -> [[Sentence]]

Makes a traceability matrix from a list of row titles, list of rows of "checked" columns, and a list of columns.

maybeChanged :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2

maybeExpanded :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2

maybeWOVerb :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2

mkEnumAbbrevList :: Integer -> Sentence -> [Sentence] -> [(Sentence, ItemType)]

Zip helper function enumerates abbreviation and zips it with list of ItemType: s - the number from which the enumeration should start from, t - the title of the list, l - the list to be enumerated

mkTableFromColumns :: [(Sentence, [Sentence])] -> ([Sentence], [[Sentence]])

Helper for making a table from a columns

noRefs :: [ItemType] -> [(ItemType, Maybe String)]

Converts lists of simple ItemTypes into a list which may be used in Contents but is not directly referable.

refineChain :: NamedIdea c => [(c, Section)] -> Sentence

Create a list in the pattern of "The __ are refined to the __". Note: Order matters!

showingCxnBw :: NamedIdea c => c -> Sentence -> Sentence

Returns the Sentence "aNamedIdea Showing the Connections Between contents"

sortBySymbol :: HasSymbol a => [a] -> [a]

Sorts a list of HasSymbols by Symbol

sortBySymbolTuple :: HasSymbol a => [(a, b)] -> [(a, b)]

Sorts a tuple list of HasSymbols by first Symbol in the tuple

substitute :: (Referable r, HasShortName r, HasSymbol r) => [r] -> Sentence

Takes a list of Referables and Symbols and outputs as a Sentence "By substituting symbols, this can be written as"

tAndDOnly :: Concept s => s -> ItemType

Helpful combinators for making Sentences for Terminologies with Definitions. term - definition

tAndDWAcc :: Concept s => s -> ItemType

Helpful combinators for making Sentences for Terminologies with Definitions. term (acc) - definition

tAndDWSym :: (Concept s, Quantity a) => s -> a -> ItemType

Helpful combinators for making Sentences for Terminologies with Definitions. term (symbol) - definition

typUncr :: HasUncertainty c => c -> Sentence

Extracts the typical uncertainty to be displayed from something that has an uncertainty

underConsidertn :: ConceptChunk -> Sentence

Returns the Sentence "The chunk under consideration is definition"

unwrap :: Maybe UnitDefn -> Sentence

Get a unit symbol if there is one

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f]

zipSentList :: [[Sentence]] -> [Sentence] -> [[Sentence]] -> [[Sentence]]

Distributes a list of Sentences by prepending individual Sentences once to an existing list of Sentences. For example: >>> zipSentList [S Hi, S Hey, S Hi] [[SHello], [SWorld], [SHello, S"World]] [[S Hi, SHello], [S Hey, SWorld], [S Hi, SHello, S"World]]

and_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a NP by combining two NamedIdeas with the word "and" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

and_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as and_ combinator, except phrase default of second term is plural instead of phrase

andRT :: (NamedIdea c, NamedIdea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Customizable and_ combinator (takes two title case capitalization rules and two NamedIdeas)

compoundNC :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Combinator for combining two NamedChunks into one. Does not preserve abbreviations

compoundNC' :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Similar to compoundNC but uses pluralNP for plural cases

compoundNC'' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> (NP -> Sentence) -> a -> b -> NamedChunk

Similar to compoundNC, except plural cases are customizable

compoundNC''' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> a -> b -> NamedChunk

compoundNCP1 :: NamedChunk -> NamedChunk -> NamedChunk

Similar to compoundNC''' but sets first parameter function to plural

compoundNCPlPh :: NamedChunk -> NamedChunk -> NamedChunk

Similar to compoundNC, except first parameter gets pluralized while second one stays singular

compoundNCPlPl :: NamedChunk -> NamedChunk -> NamedChunk

Same as compoundNC'

for :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Inserts the word "for" between the titleized versions of two terms

for' :: (NamedIdea c, Idea d) => c -> d -> Sentence

Similar to for, but uses titleized version of term 1 with the abbreviation (if it exists, phrase otherwise) for term 2

for'' :: (c -> Sentence) -> (d -> Sentence) -> c -> d -> Sentence

Similar to for, but allows one to specify the function to use on each term before inserting for. For example one could use for'' phrase plural t1 t2

of_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

of_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between them. phrase is defaulted to (phrase t1) "of" (plural t2). Plural is the same.

of_'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except plural is defaulted to (plural t1) "of" (phrase t2)

of__ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except phrase and plural default to (plural t1) "of" (phrase t2)

ofA :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of__, except combining Sentence piece is "of a"

ofN_ :: (NamedIdea c, NounPhrase d) => c -> d -> NP

Same as of_ but second argument is a NounPhrase

the :: NamedIdea t => t -> NP

Prepends "the" to a NamedIdea. Similar to the', but not titleized

the' :: NamedIdea t => t -> NP

Prepends "the" to a titleized NamedIdea

theCustom :: (t -> Sentence) -> t -> NP

A customizable version of the

with :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Case with "T1s with T2", as opposed to "T1 with T2", i.e. phrase defaults to (plural t1) "with" (phrase t2) while plural pluralizes both.

ofThe' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except combining Sentence piece is "of the"

andIts :: Sentence -> Sentence -> Sentence

Inserts the words "and its" between two Sentences

andThe :: Sentence -> Sentence -> Sentence

Inserts the words "and the" between two Sentences

fromThe :: Sentence -> Sentence -> Sentence

Inserts the words "from the" between two Sentences

inThe :: Sentence -> Sentence -> Sentence

Inserts the words "in the" between two Sentences

isExpctdToHv :: Sentence -> Sentence -> Sentence

Prepends The and inserts "is expected to have" between two Sentences

isThe :: Sentence -> Sentence -> Sentence

Inserts the words "is the" between two Sentences

ofGiv :: Sentence -> Sentence -> Sentence

Prepends "the" and inserts "of a given" between two Sentences

ofGiv' :: Sentence -> Sentence -> Sentence

Same as ofGiv, except first "the" is capitalized

ofThe :: Sentence -> Sentence -> Sentence

Inserts the words "of the" between two Sentences

the_ofThe :: Sentence -> Sentence -> Sentence

Same as ofThe, but inserts "the" at the beginning of the Sentence

the_ofThe' :: Sentence -> Sentence -> Sentence

Same as the_ofThe, except first "the" is capitalized

sOf :: Sentence -> Sentence -> Sentence

Inserts the word "of" between two Sentences

sOr :: Sentence -> Sentence -> Sentence

Inserts the word "or" between two Sentences

sVersus :: Sentence -> Sentence -> Sentence

Inserts the word "versus" between two Sentences

sAnd :: Sentence -> Sentence -> Sentence

Inserts the word "and" between two Sentences

sAre :: Sentence -> Sentence -> Sentence

Inserts the word "are" between two Sentences

sIn :: Sentence -> Sentence -> Sentence

Inserts the word "in" between two Sentences

sIs :: Sentence -> Sentence -> Sentence

Inserts the word "is" between two Sentences

toThe :: Sentence -> Sentence -> Sentence

Inserts the words "to the" between two Sentences