Utils.Drasil.Contents

Documentation

enumBullet :: Reference -> [Sentence] -> LabelledContent

Applies Enumeration, Bullet and Flat to a list.

enumBulletU :: [Sentence] -> Contents

Same as enumBullet but unlabelled.

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

Enumerates a list and applies Simple and Enumeration to it:

lb - Reference,

s - start index for the enumeration,

t - title of the list,

l - list to be enumerated.

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

Same as enumSimple but unlabelled.

eqUnR :: Expr -> Reference -> LabelledContent

Constructs LabelledContent from an expression and a reference.

eqUnR' :: Expr -> Contents

Same as eqUnR except content is unlabelled.

mkEnumSimpleD :: (Referable c, HasShortName c, Definition c) => [c] -> [Contents]

Convenience function for transforming types which are instances of the constraints Referable, HasShortName, and Definition, into Simple-type Enumerations.