Utils.Drasil.Directory

Documentation

createDirIfMissing :: Bool -> FilePath -> IO ()

Creates a directory if it does not already exist (optionally with all missing parent directories).

Implementation uses doesPathExist to check if the directory exists rather than createDirectoryIfMissing True, which would create the directory regardless of whether it exists or not, potentially leading to an error that appears in `make debug` logs.