Utils.Drasil.Document

Description

Common Doc-related functions for writting printers with a little more clarity.

Documentation

blank :: Doc

Creates a blank document with no text.

indent :: Doc -> Doc

Indents a document (by 4 spaces).

indentList :: [Doc] -> Doc

Indents a list of Docs and combines into one Doc.

contSep :: Separator

filterEmpty :: [Doc] -> [Doc]

Filter blank Docs from a list.

listToDoc :: [String] -> Doc

Merge a list of Strings into a Doc format:

e.g., `listToDoc [a,b,c,...] ~= a, b, c, ...`

type Separator = Doc

Separates document sections.