Utils.Drasil.FilePath

Documentation

data RelativeFile

A valid, relative file path with an extension in canonical form.

relativeFile :: String -> RelativeFile

Create a RelativeFile given a String that must be in canonical form, be a valid file path, contain a file extension, and be relative (not absolute); otherwise, an error is raised.

relFileToStr :: RelativeFile -> String