Utils.Drasil.Lists

Description

Functions for working with lists.

Documentation

replaceAll :: Eq a => [a] -> a -> [a] -> [a]

Replaces all elements of a target list that belong to a provided "bad" input list.

subsetOf :: Eq a => [a] -> [a] -> Bool

Checks if the first set is a subset of the second.

nubSort :: Ord a => [a] -> [a]

Sort a list, removing all duplicates

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f].

foldle :: (a -> a -> a) -> (a -> a -> a) -> a -> [a] -> a

Fold helper function that applies f to all but the last element, applies g to last element and the accumulator.

foldle1 :: (a -> a -> a) -> (a -> a -> a) -> [a] -> a

Fold helper function that applies f to all but last element, applies g to last element and accumulator without starting value, does not work for empty list.