Utils.Drasil

Documentation

enumBullet :: Reference -> [Sentence] -> LabelledContent

Applies Enumeration, Bullet and Flat to a list.

enumBulletU :: [Sentence] -> Contents

Same as enumBullet but unlabelled.

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

Enumerates a list and applies Simple and Enumeration to it:

lb - Reference,

s - start index for the enumeration,

t - title of the list,

l - list to be enumerated.

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

Same as enumSimple but unlabelled.

eqUnR :: Expr -> Reference -> LabelledContent

Constructs LabelledContent from an expression and a reference.

eqUnR' :: Expr -> Contents

Same as eqUnR except content is unlabelled.

mkEnumSimpleD :: (Referable c, HasShortName c, Definition c) => [c] -> [Contents]

Convenience function for transforming types which are instances of the constraints Referable, HasShortName, and Definition, into Simple-type Enumerations.

blank :: Doc

Creates a blank document with no text.

indent :: Doc -> Doc

Indents a document (by 4 spaces).

indentList :: [Doc] -> Doc

Indents a list of Docs and combines into one Doc.

capitalize :: String -> String

String capitalization.

stringList :: [String] -> String

Comma separated list with "and" before final item.

data EnumType

Type that helps determine enumeration method. Can use either numbers, uppercase letters, or lowercase letters.

Constructors

Numb

Upper

Lower

data WrapType

Type to help wrap a sentence with parenthesis or to add a period at the end.

Constructors

Parens

Period

data SepType

Type to help separate words with commas or semicolons.

Constructors

Comma

SemiCol

data FoldType

Type to help fold differently between listed items, or if there are options (ex. using "and" or "or" at the end of a list of words).

Constructors

List

Options

foldConstraints :: Quantity c => c -> [Constraint] -> Sentence

Helper for formatting constraints.

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements with "enumerators" in "wrappers" using foldlList.

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements separated by a "separator", ending with "and" or "or".

foldlSP :: [Sentence] -> Contents

Fold sentences then turns into content using foldlSent.

foldlSP_ :: [Sentence] -> Contents

Same as foldlSP but uses foldlSent_.

foldlSPCol :: [Sentence] -> Contents

Same as foldlSP but uses foldlSentCol.

foldlSent :: [Sentence] -> Sentence

Partial function application of foldle for sentences specifically. folds with spaces and adds "." at the end.

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period.

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon.

foldlsC :: [Sentence] -> Sentence

Folds a list of elements separated by commas, including the last element.

foldNums :: String -> [Int] -> Sentence

Parses a list of integers into a nice sentence (ie. S "1, 4-7, and 13").

numList :: String -> [Int] -> [String]

Parses a list of integers into a list of strings (ie. ["1", "4-7", "13"]).

addPercent :: Show a => a -> Sentence

Converts input to a Sentence and appends %.

bulletFlat :: [Sentence] -> ListType

Applies Bullet and Flat to a list.

bulletNested :: [Sentence] -> [ListType] -> ListType

Applies Bullets and headers to a Nested ListType. The first argument is the headers of the Nested lists.

checkValidStr :: String -> String -> Either String String

Uses an Either type to check if a String is valid - Left with error message if there is an invalid Char in String, else Right with String.

chgsStart :: (HasShortName x, Referable x) => x -> Sentence -> Sentence

Output is of the form "reference - sentence".

definedIn :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence

Takes a HasSymbol that is also Referable and outputs as a Sentence: "symbol is defined in reference."

definedIn' :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence -> Sentence

Same as definedIn, but allows for additional information to be appended to the Sentence.

definedIn'' :: (Referable r, HasShortName r) => r -> Sentence

Takes a Referable and outputs as a Sentence "defined in reference" (no HasSymbol).

displayStrConstrntsAsSet :: Quantity a => a -> [String] -> Sentence

Produces a Sentence that displays the constraints in a set {}.

displayDblConstrntsAsSet :: Quantity a => a -> [Double] -> Sentence

Produces a Sentence that displays the constraints in a set {}.

eqN :: Int -> Sentence

Prepends the word Equation to an Int.

eqnWSource :: (Referable r, HasShortName r) => Expr -> r -> Sentence

Takes an expression and a Referable and outputs as a Sentence "expression (source)".

fromReplace :: (Referable r, HasShortName r) => r -> UnitalChunk -> Sentence

Takes a Referable source and a UnitalChunk and outputs as a Sentence: "From source we can replace symbol:".

fromSource :: (Referable r, HasShortName r) => r -> Sentence

Wraps "from reference" in parentheses.

fromSources :: (Referable r, HasShortName r) => [r] -> Sentence

Similar to fromSource but takes a list of references instead of one.

fmtU :: MayHaveUnit a => Sentence -> a -> Sentence

Takes an amount as a Sentence and appends a unit to it.

follows :: (Referable r, HasShortName r) => Sentence -> r -> Sentence

Appends "following reference" to the end of a Sentence.

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in...".

itemRefToSent :: String -> Sentence -> Sentence

Makes Sentences from an item and its reference. Takes the title of reference as a String and a Sentence containing the full reference. Wraps the full reference in parenthesis.

makeListRef :: [a] -> Section -> [Sentence]

Takes a list and a reference, then generates references to match the length of the list.

makeTMatrix :: Eq a => [Sentence] -> [[a]] -> [a] -> [[Sentence]]

Makes a traceability matrix from a list of row titles, a list of rows of "checked" columns, and a list of columns.

maybeChanged :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

maybeExpanded :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

maybeWOVerb :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

mkEnumAbbrevList :: Integer -> Sentence -> [Sentence] -> [(Sentence, ItemType)]

Zip helper function enumerates abbreviations and zips it with list of ItemType:

s - the number from which the enumeration should start from (Integer),

t - the title of the list (Sentence),

l - the list to be enumerated ([Sentence]).

mkTableFromColumns :: [(Sentence, [Sentence])] -> ([Sentence], [[Sentence]])

Helper for making a table from a columns.

noRefs :: [ItemType] -> [(ItemType, Maybe String)]

Converts lists of simple ItemTypes into a list which may be used in Contents but is not directly referable.

refineChain :: NamedIdea c => [(c, Section)] -> Sentence

Create a list in the pattern of "The __ are refined to the __". Note: Order matters!

showingCxnBw :: NamedIdea c => c -> Sentence -> Sentence

Returns the Sentence "(titleize aNamedIdea) Showing the Connections Between contents".

sortBySymbol :: HasSymbol a => [a] -> [a]

Sorts a list of HasSymbols by Symbol.

sortBySymbolTuple :: HasSymbol a => [(a, b)] -> [(a, b)]

Sorts a tuple list of HasSymbols by first Symbol in the tuple.

substitute :: (Referable r, HasShortName r, HasSymbol r) => [r] -> Sentence

Takes a list of Referables and Symbols and outputs as a Sentence "By substituting symbols, this can be written as:".

tAndDOnly :: Concept s => s -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term - termDefinition".

tAndDWAcc :: Concept s => s -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term (abbreviation) - termDefinition".

tAndDWSym :: (Concept s, Quantity a) => s -> a -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term (symbol) - termDefinition".

typUncr :: HasUncertainty c => c -> Sentence

Extracts the typical uncertainty to be displayed from something that has an uncertainty.

underConsidertn :: ConceptChunk -> Sentence

Returns the Sentence "The chunk under consideration is chunkDefinition".

unwrap :: Maybe UnitDefn -> Sentence

Get a unit symbol if there is one.

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f].

zipSentList :: [[Sentence]] -> [Sentence] -> [[Sentence]] -> [[Sentence]]

Distributes a list of Sentences by prepending individual Sentences once to an existing list of Sentences.

For example:

>>> zipSentList [S "Hi", S "Hey", S "Hi"] [[S"Hello"], [S"World"], [S"Hello", S"World]]
[[S "Hi", S"Hello"], [S "Hey", S"World"], [S "Hi", S"Hello", S"World]]

fterms :: (NamedIdea c, NamedIdea d) => (NP -> NP -> t) -> c -> d -> t

Apply a binary function to the terms of two named ideas, instead of to the named ideas themselves. Ex. fterms compoundPhrase t1 t2 instead of compoundPhrase (t1 ^. term) (t2 ^. term).

and_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a NP by combining two NamedIdeas with the word "and" between their terms. Plural case is (phrase t1) "and" (plural t2).

and_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as and_ combinator, except both singular and plural case are (phrase t1) "and" (plural t2).

andRT :: (NamedIdea c, NamedIdea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Customizable and_ combinator (takes two title case capitalization rules and two NamedIdeas).

compoundNC :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Combinator for combining two NamedIdeass into a NamedChunk. Plural case only makes second term plural. See compoundPhrase for more on plural behaviour. Does not preserve abbreviations.

compoundNC' :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Similar to compoundNC but both terms are pluralized for plural case.

compoundNC'' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> (NP -> Sentence) -> a -> b -> NamedChunk

Similar to compoundNC, except plural cases are customizable.

compoundNC''' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> a -> b -> NamedChunk

Similar to compoundNC, but takes a function that is applied to the first term (eg. short or plural).

compoundNCP1 :: NamedChunk -> NamedChunk -> NamedChunk

Similar to compoundNC''' but sets first parameter function to plural.

compoundNCPlPh :: NamedChunk -> NamedChunk -> NamedChunk

Similar to compoundNC, except for plural case, first parameter gets pluralized while second one stays singular.

for :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a NP by combining two NamedIdeas with the word "for" between their terms. Plural case is (plural t1) "for" (phrase t2).

for' :: (NamedIdea c, Idea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Similar to for, but takes two functions that determine the titleCase.

of_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a NP by combining two NamedIdeas with the word "of" between their terms. Plural case is (phrase t1) "of" (plural t2).

of_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Similar to of_, but singular case is (phrase t1) "of" (plural t2). Plural case is the same.

of_'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except plural case is (plural t1) "of" (phrase t2).

of__ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except singular and plural cases are both (plural t1) "of" (phrase t2).

ofA :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_'', except combining Sentence piece is "of a".

ofA' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as ofA, except phrase case is (plural t1) "of a" (phrase t2).

ofN_ :: (NamedIdea c, NounPhrase d) => c -> d -> NP

Same as of_ but second argument is a NounPhrase.

the :: NamedIdea t => t -> NP

Prepends "the" to a NamedIdea. Similar to the', but not titleized.

the' :: NamedIdea t => t -> NP

Prepends "the" to a titleized NamedIdea.

the'' :: (t -> Sentence) -> t -> NP

A customizable version of the. The given function is applied to both singular and plural cases.

aNINP :: NamedIdea c => c -> NP

Prepends "a" to a NamedIdea (similar to the).

aNINP' :: (c -> Sentence) -> c -> NP

Customizable version of aNINP.

inThe' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a NP by combining two NamedIdeas with the words "in the" between their terms. Plural case is (plural t1) "in the" (phrase t2).

with :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Case with "T1s with T2", as opposed to "T1 with T2", i.e. singular case is (plural t1) "with" (phrase t2) while the plural case pluralizes both.

ofThe' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except combining Sentence piece is "of the". Plural case is (plural t1) ofThe (phrase t2).

the_ofThe'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as ofThe', except prepends "the".

onThe' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except combining Sentence piece is "of the".

insertStringNP :: String -> NP -> NP -> NP

Helper function that places a String in between two NPs. Plural case is (phraseNP t1) +:+ S s +:+ (pluralNP t2).

prependStringNP :: String -> NP -> NP

Helper function that prepends a String to a NP.

insertSentNP :: Sentence -> NP -> NP -> NP

Helper function that places a Sentence in between two NPs. Plural case is (phraseNP t1) +:+ s +:+ (pluralNP t2).

prependSentNP :: Sentence -> NP -> NP

Helper function that prepends a Sentence to a NP.

theNP :: NP -> NP

Prepends "the" to a NP.

theNP' :: (NP -> Sentence) -> NP -> NP

Similar to theNP, but accepts a function that determines the plural case.

aNP :: NP -> NP

Prepends "a" to a NP.

aNP' :: (NP -> Sentence) -> NP -> NP

Similar to aNP, but accepts a function that determines the plural case.

ofTheNP :: NP -> NP -> NP

Inserts "of the" between two NPs. Plural case is (phraseNP t1) +:+ "of the" +:+ (pluralNP t2).

ofTheNP' :: NP -> NP -> NP

Similar to ofTheNP, but the plural case is now (pluralNP t1) ofThe (phraseNP t2).

ofTheNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to ofTheNP, but accepts two functions for the plural case.

inTheNP :: NP -> NP -> NP

Inserts "in the" between two NPs. Plural case is (phraseNP t1) +:+ "in the" +:+ (pluralNP t2).

inTheNP' :: NP -> NP -> NP

Similar to ofTheNP, but the plural case is now (pluralNP t1) inThe (phraseNP t2).

inTheNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to ofTheNP, but accepts two functions for the plural case.

the_ofTheNP :: NP -> NP -> NP

Prepends "the" and inserts "of the". Plural case is "the" +:+ (phraseNP t1) +:+ "of the" +:+ (pluralNP t2).

the_ofTheNP' :: NP -> NP -> NP

Similar to the_ofTheNP, but the plural case is now S "the" +:+ (pluralNP t1) ofThe (phraseNP t2).

the_ofTheNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Similar to the_ofTheNP', but takes two functions for the plural case.

forNP :: NP -> NP -> NP

Inserts "for" between two NPs. Plural case is (phraseNP t1) +:+ "for" +:+ (pluralNP t2).

forNP' :: NP -> NP -> NP

Same as forNP, but plural case is now (pluralNP t1) sFor (phraseNP t2).

forNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as forNP', but takes two functions for the plural case.

ofNP :: NP -> NP -> NP

Inserts "of" between two NPs. Plural case is (phraseNP t1) +:+ "of" +:+ (pluralNP t2).

ofNP' :: NP -> NP -> NP

Same as ofNP, but plural case is now (pluralNP t1) sOf (phraseNP t2).

ofNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as ofNP, but takes two functions for the plural case.

ofNP''' :: (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as ofNP, but takes two functions for the singular case and two for the plural case.

withNP :: NP -> NP -> NP

Inserts "with" between two NPs. Plural case is (phraseNP t1) +:+ "with" +:+ (pluralNP t2).

andNP :: NP -> NP -> NP

Inserts "and" between two NPs. Plural case is (phraseNP t1) +:+ "and" +:+ (pluralNP t2).

andNP' :: NP -> NP -> NP

Same as andNP, but plural case is now (pluralNP t1) sAnd (phraseNP t2).

andNP'' :: (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as andNP, but takes two functions for the plural case.

andNP''' :: (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> (NP -> Sentence) -> NP -> NP -> NP

Same as andNP, but takes two functions for the singular case and two for the plural case.

combineNINP :: NamedIdea c => c -> NP -> NP

Helper function that combines a NamedIdea and a NP without any words in between. Plural case is phrase t1 +:+ pluralNP t2

combineNPNI :: NamedIdea c => NP -> c -> NP

Similar to combineNINP but takes in a NP first and a NamedIdea second.