Utils.Drasil

Description

Gather Drasil's utility functions and re-export for easy use.

Directory

From Utils.Drasil.Directory.

createDirIfMissing :: Bool -> FilePath -> IO ()

Creates a directory if it does not already exist (optionally with all missing parent directories).

Implementation uses doesPathExist to check if the directory exists rather than createDirectoryIfMissing True, which would create the directory regardless of whether it exists or not, potentially leading to an error that appears in `make debug` logs.

File IO

From Utils.Drasil.FileIO.

createFile :: FilePath -> String -> IO ()

Write a String to a given FilePath.

Documents

From Utils.Drasil.Document.

blank :: Doc

Creates a blank document with no text.

indent :: Doc -> Doc

Indents a document (by 4 spaces).

indentList :: [Doc] -> Doc

Indents a list of Docs and combines into one Doc.

filterEmpty :: [Doc] -> [Doc]

Filter blank Docs from a list.

listToDoc :: [String] -> Doc

Merge a list of Strings into a Doc format:

e.g., `listToDoc [a,b,c,...] ~= a, b, c, ...`

type Separator = Doc

Separates document sections.

contSep :: Separator

Language

From Utils.Drasil.English.

capitalize :: String -> String

String capitalization.

stringList :: [String] -> String

Comma separated list with "and" before final item.

FilePath

From Utils.Drasil.FilePath

data RelativeFile

A valid, relative file path with an extension in canonical form.

relativeFile :: String -> RelativeFile

Create a RelativeFile given a String that must be in canonical form, be a valid file path, contain a file extension, and be relative (not absolute); otherwise, an error is raised.

relFileToStr :: RelativeFile -> String

Lists

From Utils.Drasil.Lists. General functions involving lists.

replaceAll :: Eq a => [a] -> a -> [a] -> [a]

Replaces all elements of a target list that belong to a provided "bad" input list.

subsetOf :: Eq a => [a] -> [a] -> Bool

Checks if the first set is a subset of the second.

nubSort :: Ord a => [a] -> [a]

Sort a list, removing all duplicates

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f].

foldle :: (a -> a -> a) -> (a -> a -> a) -> a -> [a] -> a

Fold helper function that applies f to all but the last element, applies g to last element and the accumulator.

foldle1 :: (a -> a -> a) -> (a -> a -> a) -> [a] -> a

Fold helper function that applies f to all but last element, applies g to last element and accumulator without starting value, does not work for empty list.

toColumn :: [a] -> [[a]]

Convert "row" of elements into "column" of elements.

mkTable :: [a -> b] -> [a] -> [[b]]

Create a table body (not including header row) by applying the given functions to the column elements of the table rows (in order). The first argument is a list of functions to be applied (one per column). This essentially creates the rows. The second argument is a list of elements apply the functions to.

For example, mkTable [id, *5] [1,2,3] should produce a table:

| 1 |  5 |
| 2 | 10 |
| 3 | 15 |

Strings

toPlainName :: String -> String

Replace occurences of special characters (",~`-=!#$%^&*+[]\;'/|"<>? ") with underscores ("_"@).

TODO: This can probably become a bit more comprehensive, anything other than a-z, A-Z, or 0-9 could probably be replaced.

repUnd :: String -> String

Replace underscores in a string with periods (.).

TypeClasses

class HasPathAndDoc a b | a -> b where

Methods

getPath :: a -> FilePath

getDoc :: a -> b