Utils.Drasil

Description

Gather Drasil's utility functions and re-export for easy use. For now, does not include combinators (Sentence.hs, NounPhrase.hs, Concepts.hs)

Documents

From Utils.Drasil.Document.

blank :: Doc

Creates a blank document with no text.

indent :: Doc -> Doc

Indents a document (by 4 spaces).

indentList :: [Doc] -> Doc

Indents a list of Docs and combines into one Doc.

Language

From Utils.Drasil.English.

capitalize :: String -> String

String capitalization.

stringList :: [String] -> String

Comma separated list with "and" before final item.

Lists

From Utils.Drasil.Lists. General functions involving lists.

replaceAll :: Eq a => [a] -> a -> [a] -> [a]

Replaces all elements of a target list that belong to a provided "bad" input list.

subsetOf :: Eq a => [a] -> [a] -> Bool

Checks if the first set is a subset of the second.

nubSort :: Ord a => [a] -> [a]

Sort a list, removing all duplicates

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f].

foldle :: (a -> a -> a) -> (a -> a -> a) -> a -> [a] -> a

Fold helper function that applies f to all but the last element, applies g to last element and the accumulator.

foldle1 :: (a -> a -> a) -> (a -> a -> a) -> [a] -> a

Fold helper function that applies f to all but last element, applies g to last element and accumulator without starting value, does not work for empty list.

Strings

toPlainName :: String -> String

Replace occurences of special characters (",~`-=!#$%^&*+[]\;'/|"<>? ") with underscores ("_"@).

TODO: This can probably become a bit more comprehensive, anything other than a-z, A-Z, or 0-9 could probably be replaced.