Utils.Drasil

Documentation

enumBullet :: Reference -> [Sentence] -> LabelledContent

Applies Enumeration, Bullet and Flat to a list.

enumBulletU :: [Sentence] -> Contents

Same as enumBullet but unlabelled.

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

Enumerates a list and applies Simple and Enumeration to it:

lb - Reference,

s - start index for the enumeration,

t - title of the list,

l - list to be enumerated.

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

Same as enumSimple but unlabelled.

mkEnumSimpleD :: (Referable c, HasShortName c, Definition c) => [c] -> [Contents]

Convenience function for transforming types which are instances of the constraints Referable, HasShortName, and Definition, into Simple-type Enumerations.

lbldExpr :: Display c => c -> Reference -> LabelledContent

Constructs LabelledContent from an expression and a reference.

unlbldExpr :: Display c => c -> Contents

Same as eqUnR except content is unlabelled.

blank :: Doc

Creates a blank document with no text.

indent :: Doc -> Doc

Indents a document (by 4 spaces).

indentList :: [Doc] -> Doc

Indents a list of Docs and combines into one Doc.

capitalize :: String -> String

String capitalization.

stringList :: [String] -> String

Comma separated list with "and" before final item.

data EnumType

Type that helps determine enumeration method. Can use either numbers, uppercase letters, or lowercase letters.

Constructors

Numb

Upper

Lower

data WrapType

Type to help wrap a sentence with parenthesis or to add a period at the end.

Constructors

Parens

Period

data SepType

Type to help separate words with commas or semicolons.

Constructors

Comma

SemiCol

data FoldType

Type to help fold differently between listed items, or if there are options (ex. using "and" or "or" at the end of a list of words).

Constructors

List

Options

foldConstraints :: Quantity c => c -> [ConstraintE] -> Sentence

Helper for formatting constraints.

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements with "enumerators" in "wrappers" using foldlList.

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

Creates a list of elements separated by a "separator", ending with "and" or "or".

foldlSP :: [Sentence] -> Contents

Fold sentences then turns into content using foldlSent.

foldlSP_ :: [Sentence] -> Contents

Same as foldlSP but uses foldlSent_.

foldlSPCol :: [Sentence] -> Contents

Same as foldlSP but uses foldlSentCol.

foldlSent :: [Sentence] -> Sentence

Partial function application of foldle for sentences specifically. folds with spaces and adds "." at the end.

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period.

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon.

foldlsC :: [Sentence] -> Sentence

Folds a list of elements separated by commas, including the last element.

foldNums :: String -> [Int] -> Sentence

Parses a list of integers into a nice sentence (ie. S "1, 4-7, and 13").

numList :: String -> [Int] -> [String]

Parses a list of integers into a list of strings (ie. ["1", "4-7", "13"]).

replaceAll :: Eq a => [a] -> a -> [a] -> [a]

Replaces all elements of a target list that belong to a provided "bad" input list.

subsetOf :: Eq a => [a] -> [a] -> Bool

Checks if the first set is a subset of the second.

addPercent :: Show a => a -> Sentence

Converts input to a Sentence and appends %.

bulletFlat :: [Sentence] -> ListType

Applies Bullet and Flat to a list.

bulletNested :: [Sentence] -> [ListType] -> ListType

Applies Bullets and headers to a Nested ListType. The first argument is the headers of the Nested lists.

checkValidStr :: String -> String -> Either String String

Uses an Either type to check if a String is valid - Left with error message if there is an invalid Char in String, else Right with String.

chgsStart :: (HasShortName x, Referable x) => x -> Sentence -> Sentence

Output is of the form "reference - sentence".

definedIn :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence

Takes a HasSymbol that is also Referable and outputs as a Sentence: "symbol is defined in reference."

definedIn' :: (Referable r, HasShortName r, HasSymbol r) => r -> Sentence -> Sentence

Same as definedIn, but allows for additional information to be appended to the Sentence.

definedIn'' :: (Referable r, HasShortName r) => r -> Sentence

Takes a Referable and outputs as a Sentence "defined in reference" (no HasSymbol).

definedIn''' :: (HasSymbol q, HasUID q, Referable r, HasShortName r) => q -> r -> Sentence

Takes a Symbol and its Reference (does not append a period at the end!). Outputs as "symbol is defined in source".

displayStrConstrntsAsSet :: Quantity a => a -> [String] -> Sentence

Produces a Sentence that displays the constraints in a set {}.

displayDblConstrntsAsSet :: Quantity a => a -> [Double] -> Sentence

Produces a Sentence that displays the constraints in a set {}.

eqN :: Int -> Sentence

Prepends the word Equation to an Int.

eqnWSource :: (Display e, Referable r, HasShortName r) => e -> r -> Sentence

Takes an expression and a Referable and outputs as a Sentence "expression (source)".

fromReplace :: (Referable r, HasShortName r) => r -> UnitalChunk -> Sentence

Takes a Referable source and a UnitalChunk and outputs as a Sentence: "From source we can replace symbol:".

fromSource :: (Referable r, HasShortName r) => r -> Sentence

Wraps "from reference" in parentheses.

fromSources :: (Referable r, HasShortName r) => [r] -> Sentence

Similar to fromSource but takes a list of references instead of one.

fmtU :: MayHaveUnit a => Sentence -> a -> Sentence

Takes an amount as a Sentence and appends a unit to it.

follows :: (Referable r, HasShortName r) => Sentence -> r -> Sentence

Appends "following reference" to the end of a Sentence.

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in...".

itemRefToSent :: String -> Sentence -> Sentence

Makes Sentences from an item and its reference. Takes the title of reference as a String and a Sentence containing the full reference. Wraps the full reference in parenthesis.

makeListRef :: [a] -> Section -> [Sentence]

Takes a list and a reference, then generates references to match the length of the list.

makeTMatrix :: Eq a => [Sentence] -> [[a]] -> [a] -> [[Sentence]]

Makes a traceability matrix from a list of row titles, a list of rows of "checked" columns, and a list of columns.

maybeChanged :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

maybeExpanded :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

maybeWOVerb :: Sentence -> Sentence -> Sentence

Helper functions for making likely change statements. Uses form likelyFrame parameter1 _ parameter2.

mkEnumAbbrevList :: Integer -> Sentence -> [Sentence] -> [(Sentence, ItemType)]

Zip helper function enumerates abbreviations and zips it with list of ItemType:

s - the number from which the enumeration should start from (Integer),

t - the title of the list (Sentence),

l - the list to be enumerated ([Sentence]).

mkTableFromColumns :: [(Sentence, [Sentence])] -> ([Sentence], [[Sentence]])

Helper for making a table from a columns.

noRefs :: [ItemType] -> [(ItemType, Maybe String)]

Converts lists of simple ItemTypes into a list which may be used in Contents but is not directly referable.

refineChain :: NamedIdea c => [(c, Section)] -> Sentence

Create a list in the pattern of "The __ are refined to the __". Note: Order matters!

showingCxnBw :: NamedIdea c => c -> Sentence -> Sentence

Returns the Sentence "(titleize aNamedIdea) Showing the Connections Between contents".

sortBySymbol :: HasSymbol a => [a] -> [a]

Sorts a list of HasSymbols by Symbol.

sortBySymbolTuple :: HasSymbol a => [(a, b)] -> [(a, b)]

Sorts a tuple list of HasSymbols by first Symbol in the tuple.

substitute :: (Referable r, HasShortName r, HasSymbol r) => [r] -> Sentence

Takes a list of Referables and Symbols and outputs as a Sentence "By substituting symbols, this can be written as:".

tAndDOnly :: Concept s => s -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term - termDefinition".

tAndDWAcc :: Concept s => s -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term (abbreviation) - termDefinition".

tAndDWSym :: (Concept s, Quantity a) => s -> a -> ItemType

Helpful combinators for making Sentences into Terminologies with Definitions. Returns of the form: "term (symbol) - termDefinition".

typUncr :: HasUncertainty c => c -> Sentence

Extracts the typical uncertainty to be displayed from something that has an uncertainty.

underConsidertn :: ConceptChunk -> Sentence

Returns the Sentence "The chunk under consideration is chunkDefinition".

unwrap :: Maybe UnitDefn -> Sentence

Get a unit symbol if there is one.

weave :: [[a]] -> [a]

Interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f].

zipSentList :: [[Sentence]] -> [Sentence] -> [[Sentence]] -> [[Sentence]]

Distributes a list of Sentences by prepending individual Sentences once to an existing list of Sentences.

For example:

>>> zipSentList [S "Hi", S "Hey", S "Hi"] [[S"Hello"], [S"World"], [S"Hello", S"World]]
[[S "Hi", S"Hello"], [S "Hey", S"World"], [S "Hi", S"Hello", S"World]]

fterms :: (NamedIdea c, NamedIdea d) => (NP -> NP -> t) -> c -> d -> t

Apply a binary function to the terms of two named ideas, instead of to the named ideas themselves. Ex. fterms compoundPhrase t1 t2 instead of compoundPhrase (t1 ^. term) (t2 ^. term).