Utils.Drasil

Documentation

data EnumType

Constructors

Numb

Upper

Lower

data WrapType

Constructors

Parens

Period

data SepType

Constructors

Comma

SemiCol

data FoldType

Constructors

List

Options

foldConstraints :: Quantity c => c -> [Constraint] -> Sentence

helper for formatting constraints

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

creates an list of elements with "enumerators" in "wrappers" using foldlList

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

creates a list of elements separated by a "separator", ending with "and" or "or"

foldlSP :: [Sentence] -> Contents

fold sentences then turns into content

foldlSP_ :: [Sentence] -> Contents

foldlSPCol :: [Sentence] -> Contents

foldlSent :: [Sentence] -> Sentence

partial function application of foldle for sentences specifically

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon

foldlsC :: [Sentence] -> Sentence

creates a list of elements separated by commas, including the last element

addPercent :: Show a => a -> Sentence

outputs sentence with % attached to it

bulletFlat :: [Sentence] -> ListType

bulletFlat applies Bullet and Flat to a list.

bulletNested :: [Sentence] -> [ListType] -> ListType

bulletNested applies Bullets and headers to a Nested ListType. t - Headers of the Nested lists. l - Lists of ListType.

chgsStart :: (HasShortName x, Referable x) => x -> Sentence -> Sentence

displayConstrntsAsSet :: Quantity a => a -> [String] -> Sentence

enumBullet :: Reference -> [Sentence] -> LabelledContent

enumBullet apply Enumeration, Bullet and Flat to a list

enumBulletU :: [Sentence] -> Contents

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

enumSimple enumerates a list and applies simple and enumeration to it s - start index for the enumeration t - title of the list l - list to be enumerated

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

eqN :: Int -> Sentence

eqUnR :: Expr -> Reference -> LabelledContent

eqUnR' :: Expr -> Contents

fmtU :: MayHaveUnit a => Sentence -> a -> Sentence

takes a amount and adds a unit to it n - sentenc representing an amount u - unit we want to attach to amount

follows :: (Referable r, HasShortName r) => Sentence -> r -> Sentence

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in..."

itemRefToSent :: String -> Sentence -> Sentence

makes sentences from an item and its reference a - String title of reference b - Sentence containing the full reference

makeListRef :: [a] -> Section -> [Sentence]

makeListRef takes a list and a reference and generates references to match the length of the list l - list whos length is to be matched

makeTMatrix :: Eq a => [Sentence] -> [[a]] -> [a] -> [[Sentence]]

makes a traceability matrix from list of row title, list of rows of "checked" columns, and a list of columns.

maybeChanged :: Sentence -> Sentence -> Sentence

maybeExpanded :: Sentence -> Sentence -> Sentence

maybeWOVerb :: Sentence -> Sentence -> Sentence

mkEnumAbbrevList :: Integer -> Sentence -> [Sentence] -> [(Sentence, ItemType)]

zip helper function enumerates abbreviation and zips it with list of itemtype s - the number from which the enumeration should start from t - the title of the list l - the list to be enumerated

mkTableFromColumns :: [(Sentence, [Sentence])] -> ([Sentence], [[Sentence]])

Helper for making a table from a columns

noRefs :: [ItemType] -> [(ItemType, Maybe String)]

noRefs converts lists of simple ItemTypes into a lists which may be used in Contents but not directly referable.

noRefsLT :: [(Sentence, ItemType)] -> [ListTuple]

noRefsLT converts lists of tuples containing a title and ItemType into a ListTuple which can be used with Contents but not directly referable.

refineChain :: NamedIdea c => [(c, Section)] -> Sentence

Create a list in the pattern of "The are refined to the ". Note: Order matters!

showingCxnBw :: NamedIdea c => c -> Sentence -> Sentence

sortBySymbol :: HasSymbol a => [a] -> [a]

sortBySymbolTuple :: HasSymbol a => [(a, b)] -> [(a, b)]

tAndDOnly :: Concept s => s -> ItemType

tAndDWAcc :: Concept s => s -> ItemType

helpful combinators for making Sentences for Terminologies with Definitions term (acc) - definition

tAndDWSym :: (Concept s, Quantity a) => s -> a -> ItemType

typUncr :: HasUncertainty c => c -> Sentence

extracts the typical uncertainty to be displayed from something that has an uncertainty

underConsidertn :: ConceptChunk -> Sentence

unwrap :: Maybe UnitDefn -> Sentence

get a unit symbol if there is one

weave :: [[a]] -> [a]

interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f]

zipSentList :: [[Sentence]] -> [Sentence] -> [[Sentence]] -> [[Sentence]]

appends a sentence to the front of a list of list of sentences

and_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates an NP by combining two NamedIdeas with the word "and" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

and_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as and_ combinator, except phrase default of second term is plural instead of phrase

andRT :: (NamedIdea c, NamedIdea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Customizable and combinator

compoundNC :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Combinator for combining two NamedChunks into one. Does not preserve abbreviations

compoundNC' :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

compoundNC'' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> (NP -> Sentence) -> a -> b -> NamedChunk

compoundNC''' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> a -> b -> NamedChunk

compoundNCP1 :: NamedChunk -> NamedChunk -> NamedChunk

compoundNCPlPh :: NamedChunk -> NamedChunk -> NamedChunk

compoundNCPlPl :: NamedChunk -> NamedChunk -> NamedChunk

for :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Inserts the word "for" between the titleized versions of two terms

for' :: (NamedIdea c, Idea d) => c -> d -> Sentence

Similar to for, but uses titleized version of term 1 with the abbreviation (if it exists, phrase otherwise) for term 2

for'' :: (c -> Sentence) -> (d -> Sentence) -> c -> d -> Sentence

Similar to for, but allows one to specify the function to use on each term before inserting for. For example one could use for'' phrase plural t1 t2

of_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

of_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between them. phrase is defaulted to (phrase t1) "of" (plural t2). Plural is the same.

of_'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except plural default of second term is phrase

of__ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except phrase default of first term is plural instead of phrase

ofA :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of__, except combining Sentence piece is "of a"

ofN_ :: (NamedIdea c, NounPhrase d) => c -> d -> NP

the :: NamedIdea t => t -> NP

the' :: NamedIdea t => t -> NP

theCustom :: (t -> Sentence) -> t -> NP

with :: (NamedIdea c, NamedIdea d) => c -> d -> NP

andIts :: Sentence -> Sentence -> Sentence

andThe :: Sentence -> Sentence -> Sentence

isExpctdToHv :: Sentence -> Sentence -> Sentence

isThe :: Sentence -> Sentence -> Sentence

ofGiv :: Sentence -> Sentence -> Sentence

ofGiv' :: Sentence -> Sentence -> Sentence

ofThe :: Sentence -> Sentence -> Sentence

ofThe' :: Sentence -> Sentence -> Sentence

sOf :: Sentence -> Sentence -> Sentence

sOr :: Sentence -> Sentence -> Sentence

sVersus :: Sentence -> Sentence -> Sentence

sAnd :: Sentence -> Sentence -> Sentence

sAre :: Sentence -> Sentence -> Sentence

sIn :: Sentence -> Sentence -> Sentence

sIs :: Sentence -> Sentence -> Sentence

toThe :: Sentence -> Sentence -> Sentence