Utils.Drasil

Documentation

data EnumType

Constructors

Numb

Upper

Lower

data WrapType

Constructors

Parens

Period

data SepType

Constructors

Comma

SemiCol

data FoldType

Constructors

List

Options

foldConstraints :: Quantity c => c -> [Constraint] -> Sentence

helper for formatting constraints

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

creates an list of elements with "enumerators" in "wrappers" using foldlList

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

creates a list of elements separated by a "separator", ending with "and" or "or"

foldlSP :: [Sentence] -> Contents

fold sentences then turns into content

foldlSP_ :: [Sentence] -> Contents

foldlSPCol :: [Sentence] -> Contents

foldlSent :: [Sentence] -> Sentence

partial function application of foldle for sentences specifically

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon

foldlsC :: [Sentence] -> Sentence

creates a list of elements separated by commas, including the last element

and_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates an NP by combining two NamedIdeas with the word "and" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

and_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as and_ combinator, except phrase default of second term is plural instead of phrase

andRT :: (NamedIdea c, NamedIdea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Customizable and combinator

compoundNC :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Combinator for combining two NamedChunks into one. Does not preserve abbreviations

compoundNC' :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

compoundNC'' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> (NP -> Sentence) -> a -> b -> NamedChunk

compoundNC''' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> a -> b -> NamedChunk

compoundNCP1 :: NamedChunk -> NamedChunk -> NamedChunk

compoundNCPlPh :: NamedChunk -> NamedChunk -> NamedChunk

compoundNCPlPl :: NamedChunk -> NamedChunk -> NamedChunk

for :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Inserts the word "for" between the titleized versions of two terms

for' :: (NamedIdea c, Idea d) => c -> d -> Sentence

Similar to for, but uses titleized version of term 1 with the abbreviation (if it exists, phrase otherwise) for term 2

for'' :: (c -> Sentence) -> (d -> Sentence) -> c -> d -> Sentence

Similar to for, but allows one to specify the function to use on each term before inserting for. For example one could use for'' phrase plural t1 t2

of_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

of_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between them. phrase is defaulted to (phrase t1) "of" (plural t2). Plural is the same.

of_'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except plural default of second term is phrase

of__ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except phrase default of first term is plural instead of phrase

ofA :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of__, except combining Sentence piece is "of a"

ofN_ :: (NamedIdea c, NounPhrase d) => c -> d -> NP

the :: NamedIdea t => t -> NP

the' :: NamedIdea t => t -> NP

theCustom :: (t -> Sentence) -> t -> NP

with :: (NamedIdea c, NamedIdea d) => c -> d -> NP

andIts :: Sentence -> Sentence -> Sentence

andThe :: Sentence -> Sentence -> Sentence

isExpctdToHv :: Sentence -> Sentence -> Sentence

isThe :: Sentence -> Sentence -> Sentence

ofGiv :: Sentence -> Sentence -> Sentence

ofGiv' :: Sentence -> Sentence -> Sentence

ofThe :: Sentence -> Sentence -> Sentence

ofThe' :: Sentence -> Sentence -> Sentence

sOf :: Sentence -> Sentence -> Sentence

sOr :: Sentence -> Sentence -> Sentence

sVersus :: Sentence -> Sentence -> Sentence

sAnd :: Sentence -> Sentence -> Sentence

sAre :: Sentence -> Sentence -> Sentence

sIn :: Sentence -> Sentence -> Sentence

sIs :: Sentence -> Sentence -> Sentence

toThe :: Sentence -> Sentence -> Sentence