Utils.Drasil.Lists

Description

Functions for working with lists.

Documentation

replaceAll :: Eq a => [a] -> a -> [a] -> [a]

Replaces all elements of a target list that belong to a provided "bad" input list.

subsetOf :: Eq a => [a] -> [a] -> Bool

Checks if the first set is a subset of the second.